Spam a lot of messages with the NUL character
Cleanup plugin syntax code-style wise

Remove setting covered by general inventory update behavior

Move plugin so inventory can pick it up
Move inventory file too
