+++ ansible.cfg
@@ -1,3 +1,2 @@
 [defaults]
 deprecation_warnings=false
-collections_paths=collections
