+++ inventories/bad.yml
